+++ lib/core/localization/app_translations.dart
@@ -1,0 +1,47 @@
+import 'package:get/get.dart';
+
+class AppTranslations extends Translations {
+  @override
+  Map<String, Map<String, String>> get keys => {
+        'en_US': {
+          'where_are_you_going': 'Where are you going?',
+          'save_to_favorites': 'Save to Favorites',
+          'position_destination': 'Position destination at center pin',
+          'confirm_location': 'Confirm Location',
+          'start_tracking_here': 'Start Tracking Here',
+          'delete_location': 'Delete Location',
+          'location_saved': 'Location saved as @label',
+          'saved': 'Saved',
+          'deleted': 'Deleted',
+          'location_removed': '@label removed from saved locations.',
+          'search_destination': 'Search for a destination',
+          'start_tracking': 'Start Tracking',
+          'stop_tracking': 'Stop Tracking',
+          'arriving_in': 'Arriving in',
+          'distance': 'Distance',
+          'speed': 'Speed',
+          'eta': 'ETA',
+          'cancel': 'Cancel',
+          'saved_locations': 'Saved Locations',
+          'results': 'Results',
+          'test_nearby': 'Test Nearby',
+          'tag_location': 'Tag this location',
+          'location_name': 'Location Name (e.g. Home)',
+          'icon': 'Icon',
+          'save': 'Save',
+          'waiting_for_movement': 'Waiting for movement...',
+          'less_than_1_min': '< 1 min',
+          'mins': '@mins min',
+          'hours_mins': '@hours h @mins m',
+          'moving_eta': 'Moving · ETA @eta',
+          'moving': 'Moving',
+          'km_to_destination': '@km km to @destination',
+          'alert_when_within': 'Alert when within @distance m',
+          'km_h': 'km/h',
+          'km_remaining': '@km km remaining',
+          'stop_alert': 'Stop Alert',
+          'destination_placeholder': 'Destination',
+        },
+      };
+}
+
