+++ lib/modules/home/views/widgets/loading_screen.dart
@@ -1,0 +1,117 @@
+import 'package:flutter/material.dart';
+import 'package:get/get.dart';
+import 'package:locami/modules/home/controllers/home_controller.dart';
+
+class LoadingScreen extends StatelessWidget {
+  final bool isDark;
+  final Color accentColor;
+
+  const LoadingScreen({
+    super.key,
+    required this.isDark,
+    required this.accentColor,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    final controller = Get.find<HomeController>();
+
+    return Container(
+      color: isDark ? const Color(0xFF0A0A0A) : const Color(0xFFF1F3F5),
+      child: Center(
+        child: Column(
+          mainAxisSize: MainAxisSize.min,
+          children: [
+            SizedBox(
+              width: 120,
+              height: 120,
+              child: Stack(
+                alignment: Alignment.center,
+                children: [
+                  TweenAnimationBuilder<double>(
+                    tween: Tween(begin: 0.0, end: 1.0),
+                    duration: const Duration(milliseconds: 1500),
+                    builder: (context, value, child) {
+                      return Container(
+                        width: 100 + (20 * value),
+                        height: 100 + (20 * value),
+                        decoration: BoxDecoration(
+                          shape: BoxShape.circle,
+                          border: Border.all(
+                            color: accentColor.withValues(alpha: 0.3 * (1 - value)),
+                            width: 2,
+                          ),
+                        ),
+                      );
+                    },
+                    onEnd: () {},
+                  ),
+                  Container(
+                    width: 80,
+                    height: 80,
+                    decoration: BoxDecoration(
+                      shape: BoxShape.circle,
+                      border: Border.all(
+                        color: accentColor.withValues(alpha: 0.2),
+                        width: 1.5,
+                      ),
+                    ),
+                  ),
+                  Container(
+                    width: 20,
+                    height: 20,
+                    decoration: BoxDecoration(
+                      color: accentColor,
+                      shape: BoxShape.circle,
+                      boxShadow: [
+                        BoxShadow(
+                          color: accentColor.withValues(alpha: 0.5),
+                          blurRadius: 20,
+                          spreadRadius: 5,
+                        ),
+                      ],
+                    ),
+                  ),
+                ],
+              ),
+            ),
+            const SizedBox(height: 48),
+            Text(
+              'LOCAMI',
+              style: TextStyle(
+                color: isDark ? Colors.white : Colors.black87,
+                fontSize: 28,
+                fontWeight: FontWeight.w800,
+                letterSpacing: 6,
+              ),
+            ),
+            const SizedBox(height: 32),
+            Obx(() => AnimatedSwitcher(
+              duration: const Duration(milliseconds: 300),
+              child: Text(
+                controller.initStatus.value,
+                key: ValueKey(controller.initStatus.value),
+                textAlign: TextAlign.center,
+                style: TextStyle(
+                  color: isDark ? Colors.white54 : Colors.black45,
+                  fontSize: 14,
+                  fontWeight: FontWeight.w500,
+                  letterSpacing: 0.5,
+                ),
+              ),
+            )),
+            const SizedBox(height: 24),
+            SizedBox(
+              width: 32,
+              height: 32,
+              child: CircularProgressIndicator(
+                strokeWidth: 2,
+                color: accentColor.withValues(alpha: 0.6),
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
